+++ tests/add_single_task.spec.js
@@ -1,4 +1,5 @@
-import { expect, test, waitPageLoading } from './base_test.js';
+import { expect, test } from '@playwright/test';
+import { waitPageLoading } from './utils.js';
 import { fileURLToPath } from 'url';
 import path from 'path';
 
@@ -10,7 +11,7 @@ test('Add single tasks', async ({ page }) => {
 	await waitPageLoading(page);
 
 	await test.step('Select "Add a single task" form', async () => {
-		await page.getByLabel('Single task').click({ force: true });
+		await page.getByText('Single task').click();
 	});
 
 	const createBtn = page.getByRole('button', { name: /^Create$/ });