+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 *Note: Numbers like (\#123) point to closed Pull Requests on the fractal-web repository.*
 
+# Unreleased
+
+* Removed coverage setup from Playwright tests and CI (\#399).
+
 # 0.8.2
 
 * Improvements on worflow task form (\#393):